Separated modules and scripts
