+++ src/scripts/real_scripts/CSM_add_network.py
