+++ src/scripts/real_scripts/CSM_add_network_from_txt.py
