+++ src/scripts/real_scripts/CSM_setup_home_network.py
