+++ src/scripts/real_scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Scripts to be called externally."""
